lang-chain/Azure/requirements.txt
python-dotenv
langchain
langchain-community
sentence_transformers
azure-search-documents
azure-identity
beautifulsoup4